pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "persona-memory"
version = "0.1.0"
description = "Evolving working memory for Alexa and Jarvis personas with Gemini integration"
authors = [{ name = "User" }]
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.8",
  "httpx>=0.27",
  "tenacity>=9.0",
  "sqlite-utils>=3.36",
  "python-dotenv>=1.0",
  "uvloop>=0.20; platform_system == 'Linux'",
]

[project.scripts]
persona-memory = "app.main:main"
